COMP: Fix sign-compare warning in qSlicerTransformPreviewWidget.cxx
See #143

+++ FiducialRegistrationWizard/Widgets/qSlicerTransformPreviewWidget.cxx
@@ -317,7 +317,7 @@ void qSlicerTransformPreviewWidget::clearPreviewNodes()
 {
   Q_D(qSlicerTransformPreviewWidget);
 
-  for ( int i = 0; i < d->PreviewNodes.size(); i++ )
+  for ( size_t i = 0; i < d->PreviewNodes.size(); i++ )
   {
     this->mrmlScene()->RemoveNode( d->PreviewNodes.at(i) );
   }
